Sweetalert, fuck yeah

+++ source/views/userpage/userpage.js
@@ -59,12 +59,27 @@ angular.module('randomEncounter')
 		}
 	}
 	$scope.removeInterest = function(interestId){
-		InterestService.removeInterestFromSchema(interestId, $rootScope.authentication.userInfo._id)
-		.then(function(res){
-			$scope.getInterests($rootScope.authentication.userInfo._id);
-		})
-	}
-})
+		swal({
+			title: "Are you sure?",
+			text: "You will not be able to recover this imaginary file!",
+			type: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#DD6B55",
+			confirmButtonText: "Yes, delete it!",
+			cancelButtonText: "No, cancel plx!",
+			closeOnConfirm: false,
+			closeOnCancel: false },
+			function(isConfirm){
+				if (isConfirm) {
+					InterestService.removeInterestFromSchema(interestId, $rootScope.authentication.userInfo._id)
+					.then(function(res){
+						$scope.getInterests($rootScope.authentication.userInfo._id);
+						swal("Deleted!", "Your imaginary file has been deleted.", "success");
+					})
+				} else {
+					swal("Cancelled", "Your imaginary file is safe :)", "error");   } });
+				}
+			})
 
 
 
@@ -83,4 +98,4 @@ angular.module('randomEncounter')
 
 
 
-//
+			//
